zfs_is_mounted fix for com.apple.devdisk
Correctly find mounted datasets even if their com.apple.devdisk property
is enabled. This is done by searching for the dataset name and the BSD Name
found in the IORegistry, if available.
Currently, this search is added in getmntany, at a very low level, so it
applies to all users.